Move MainWindow object to its own file
